+++ Editor/Config/ModuleConfig/ED_MainModuleDirectoryStructure.cs
@@ -106,13 +106,6 @@ namespace FlowIoC.Editor.Config.ModuleConfig
                             },
                             new FolderEVO
                             {
-                                FolderName = "Functions",
-                                Type = FolderEVO.FolderType.Folder,
-                                IsMandatory = true,
-                                IsNamespaceProvider = true
-                            },
-                            new FolderEVO
-                            {
                                 FolderName = "Services",
                                 Type = FolderEVO.FolderType.Folder,
                                 IsMandatory = false,
@@ -282,6 +275,7 @@ namespace FlowIoC.Editor.Config.ModuleConfig
             bool healed = config.EnsureSharedBranch(settings);
             healed |= config.EnsurePublicSignalsFolder(settings);
             healed |= config.RemoveRetiredFolderTypes();
+            healed |= config.RemoveRetiredFolderNames();
             healed |= config.MakeFolderOptional("Scriptables");
 
             if (healed)
@@ -336,7 +330,6 @@ namespace FlowIoC.Editor.Config.ModuleConfig
                             FolderEVO.FolderType.Signals, null, false, true),
                         CreateFolder(codeGenSettings.DirectoryStructureConfigMap[FolderEVO.FolderType.ViewsAndMediators],
                             FolderEVO.FolderType.ViewsAndMediators, null, true),
-                        CreateFolder("Functions", FolderEVO.FolderType.Folder, null, true, isNamespaceProvider: true),
                         CreateFolder(codeGenSettings.DirectoryStructureConfigMap[FolderEVO.FolderType.Services], FolderEVO.FolderType.Services,
                             null, false, true, isNamespaceProvider: true),
                         CreateFolder(codeGenSettings.FolderNameFor(FolderEVO.FolderType.Systems, "Systems"), FolderEVO.FolderType.Systems,