+++ Editor/Config/ModuleConfig/ED_TestModuleDirectoryStructure.cs
@@ -106,13 +106,6 @@ namespace FlowIoC.Editor.Config.ModuleConfig
                             },
                             new FolderEVO
                             {
-                                FolderName = "Functions",
-                                Type = FolderEVO.FolderType.Folder,
-                                IsMandatory = true,
-                                IsNamespaceProvider = true
-                            },
-                            new FolderEVO
-                            {
                                 FolderName = "Services",
                                 Type = FolderEVO.FolderType.Services,
                                 IsMandatory = false,
@@ -212,6 +205,7 @@ namespace FlowIoC.Editor.Config.ModuleConfig
             }
 
             bool healed = config.RemoveRetiredFolderTypes();
+            healed |= config.RemoveRetiredFolderNames();
             healed |= config.MakeFolderOptional("Scriptables");
 
             if (healed)
@@ -264,7 +258,6 @@ namespace FlowIoC.Editor.Config.ModuleConfig
                             FolderEVO.FolderType.Signals, null, false, true, isNamespaceProvider: true),
                         CreateFolder(codeGenSettings.DirectoryStructureConfigMap[FolderEVO.FolderType.ViewsAndMediators],
                             FolderEVO.FolderType.ViewsAndMediators, null, true, isNamespaceProvider: true),
-                        CreateFolder("Functions", FolderEVO.FolderType.Folder, null, true, isNamespaceProvider: true),
                         CreateFolder(codeGenSettings.DirectoryStructureConfigMap[FolderEVO.FolderType.Services], FolderEVO.FolderType.Services,
                             null, false, true, isNamespaceProvider: true),
                         CreateFolder("Constants", FolderEVO.FolderType.Folder, null, false, true, isNamespaceProvider: true),